PB-1434 : refactoring change to LabelWidget
