+++ app/stac_api/widgets/label_widget.py
